Refactor products sort
