+++ saleor/graphql/product/types/__init__.py
@@ -7,7 +7,6 @@ from .products import (
     MoveProductInput,
     Product,
     ProductImage,
-    ProductOrder,
     ProductType,
     ProductVariant,
 )
